refactor dumbtorrent, it is done now
