+++ core/dumbtorrent/dumbtorrent.go
@@ -1,4 +1,4 @@
-package downloader
+package dumbtorrent
 
 import (
 	"fmt"
